Hilo funcionando del envio de correo y la busqueda por el estado al crear la cita, tambien propiedades extras de citas mappeadas

+++ BussinessLayer/DTOs/ModuloCitas/CtaState/CtaStateResponse.cs
@@ -6,8 +6,8 @@
         public string? Description { get; set; }
         public bool IsClosure { get; set; } = false;
         public DateTime CreationDate { get; set; } = DateTime.Now;
-        public int EmailTemplateIdIn { get; set; }
-        public int EmailTemplateIdOut { get; set; }
+        public long EmailTemplateIdIn { get; set; }
+        public long EmailTemplateIdOut { get; set; }
         public long CompanyId { get; set; }
         public int AreaId { get; set; }
         public bool IsDefault { get; set; }
