+++ DataLayer/Models/ModuloCitas/CtaState.cs
@@ -10,8 +10,8 @@ namespace DataLayer.Models.Modulo_Citas
         public string? Description { get; set; }
         public bool IsClosure { get; set; } = false;
         public DateTime CreationDate { get; set; } = DateTime.Now;
-        public int EmailTemplateIdIn { get; set; }
-        public int EmailTemplateIdOut { get; set; }
+        public long EmailTemplateIdIn { get; set; }
+        public long EmailTemplateIdOut { get; set; }
         public long CompanyId { get; set; }
         public int AreaId { get; set; }
         public bool IsDefault { get; set; }
